requirements.txt
pandas==1.4.1
joblib>=1.2.0
numpy==1.21.6
pillow
scikit-learn==1.0.2
streamlit==1.13.0
tensorflow==2.10.0
